Fix double toast bug
