+++ src/extensions/core/electronAdapter.ts
@@ -1,8 +1,10 @@
+import log from 'loglevel'
+
 import { PYTHON_MIRROR } from '@/constants/uvMirrors'
 import { t } from '@/i18n'
 import { app } from '@/scripts/app'
 import { useDialogService } from '@/services/dialogService'
-import { useUpdateService } from '@/services/updateService'
+import { useToastStore } from '@/stores/toastStore'
 import { useWorkflowStore } from '@/stores/workflowStore'
 import { electronAPI as getElectronAPI, isElectron } from '@/utils/envUtil'
 import { checkMirrorReachable } from '@/utils/networkUtil'
@@ -13,7 +15,7 @@ import { checkMirrorReachable } from '@/utils/networkUtil'
   const electronAPI = getElectronAPI()
   const desktopAppVersion = await electronAPI.getElectronVersion()
   const workflowStore = useWorkflowStore()
-  const updateService = useUpdateService()
+  const toastStore = useToastStore()
 
   const onChangeRestartApp = (newValue: string, oldValue: string) => {
     // Add a delay to allow changes to take effect before restarting.
@@ -165,7 +167,46 @@ import { checkMirrorReachable } from '@/utils/networkUtil'
         label: 'Check for Updates',
         icon: 'pi pi-sync',
         async function() {
-          await updateService.checkForUpdates()
+          try {
+            const updateInfo = await electronAPI.checkForUpdates({
+              disableUpdateReadyAction: true
+            })
+
+            if (!updateInfo.isUpdateAvailable) {
+              toastStore.add({
+                severity: 'info',
+                summary: t('desktopUpdate.noUpdateFound')
+              })
+              return
+            }
+
+            const proceed = await useDialogService().confirm({
+              title: t('desktopUpdate.updateFoundTitle', {
+                version: updateInfo.version
+              }),
+              message: t('desktopUpdate.updateAvailableMessage'),
+              type: 'default'
+            })
+            if (proceed) {
+              try {
+                electronAPI.restartAndInstall()
+              } catch (error) {
+                log.error('Error installing update:', error)
+                toastStore.add({
+                  severity: 'error',
+                  summary: t('g.error'),
+                  detail: t('desktopUpdate.errorInstallingUpdate')
+                })
+              }
+            }
+          } catch (error) {
+            log.error('Error checking for updates:', error)
+            toastStore.add({
+              severity: 'error',
+              summary: t('g.error'),
+              detail: t('desktopUpdate.errorCheckingUpdate')
+            })
+          }
         }
       },
       {
